requirements.txt
bar-chart-race==0.1.0
calplot==0.1.7.5
contourpy==1.3.2
cramjam==2.10.0
cycler==0.12.1
fastparquet==2024.11.0
fonttools==4.58.0
fsspec==2025.3.2
imageio==2.37.0
kiwisolver==1.4.8
matplotlib==3.10.3
narwhals==1.39.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
plotly==6.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
seaborn==0.13.2
six==1.17.0
tzdata==2025.2
wordcloud==1.9.4
